applications: cup2023: add command to test central lift
The application shell should be soon removed. But for now, no interface
allows to test motor actuators, so add the lift command for testing.
This shell menu will be removed in a future commit.
