+++ applications/cup2023/app.cpp
@@ -1,11 +1,11 @@
 /* Project includes */
 #include "app.hpp"
 #include "app_conf.hpp"
-#include "app_arm_shell.hpp"
+#include "app_actuators_shell.hpp"
 #include "platform.hpp"
 
 void app_init(void)
 {
     /* Init shell commands */
-    cogip::app::arms::shell_init();
+    cogip::app::actuators::shell_init();
 }
